govern_cpi

Module cpi

Modules

accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is an AccountInfo. This is useful for CPI.

Structs

Return

Functions

activate_proposal

cancel_proposal

create_governor

create_proposal

create_proposal_meta

new_vote

queue_proposal

set_electorate

set_governance_params

set_vote