govern_cpi::cpi

Module accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is an AccountInfo. This is useful for CPI.

Structs

ActivateProposal

Generated CPI struct of the accounts for ActivateProposal.

CancelProposal

Generated CPI struct of the accounts for CancelProposal.

CreateGovernor

Generated CPI struct of the accounts for CreateGovernor.

CreateProposal

Generated CPI struct of the accounts for CreateProposal.

CreateProposalMeta

Generated CPI struct of the accounts for CreateProposalMeta.

NewVote

Generated CPI struct of the accounts for NewVote.

QueueProposal

Generated CPI struct of the accounts for QueueProposal.

SetElectorate

Generated CPI struct of the accounts for SetElectorate.

SetGovernanceParams

Generated CPI struct of the accounts for SetGovernanceParams.

SetVote

Generated CPI struct of the accounts for SetVote.